data-engineering-zoomcamp/pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "data-engineering-zoomcamp"
version = "0.1.0"
description = "Data engineering project for zoomcamp"
dependencies = [
    "click",
    "pandas",
    "sqlalchemy",
    "tqdm",
    "psycopg2-binary",
    "pyarrow"
]

[tool.setuptools]
packages = ["src"]